+++ src/Models/Media.php
@@ -56,6 +56,17 @@ class Media extends Model
         return $this->getUrl(MediaHub::getThumbnailConversionName());
     }
 
+    public function getExtension()
+    {
+        $extension = (new \Symfony\Component\Mime\MimeTypes)->getExtensions($this->mime_type)[0];
+        if (!$extension) $extension = str($this->file_name)->afterLast('.')->toString();
+
+        // Fix issue with Imagick not recognizing "tif" as a valid format
+        if ($extension === 'tif') $extension = 'tiff';
+
+        return $extension;
+    }
+
     public function formatForNova()
     {
         return [